Fix Windows build: move decodeRunes test to unix-only file
